+++ src/objects/date.c
@@ -120,11 +120,11 @@ void mud_date_adjust(mud_date_t * date, mud_int_t adjust) {
 mud_int_t mud_date_to_timestamp(mud_date_t * date) {
   struct tm t = { 0 };
   t.tm_isdst = -1;
-  t.tm_year = date->year - 1900;
-  t.tm_mon = date->mon - 1;
-  t.tm_mday = date->mday;
-  t.tm_hour = date->hour;
-  t.tm_min = date->sec / 60;
+  t.tm_year = (int)date->year - 1900;
+  t.tm_mon = (int)date->mon - 1;
+  t.tm_mday = (int)date->mday;
+  t.tm_hour = (int)date->hour;
+  t.tm_min = (int)date->sec / 60;
   t.tm_sec = date->sec % 60;
   return mktime(&t);
 }
